erofs-utils: tests: add test for xattrs in random layouts
Generate and check random xattrs.  Set xattrs on tested file as much
as possible until the number or size of xattrs has reached the upper
limit for single file, e.g. maximum one block size for xattrs for
single file in ext4.

The same xattr may be shared among multiple tested files or not.
Thus the layout of mixed inline and shared xattrs is tested.

Signed-off-by: Jingbo Xu <jefflexu@linux.alibaba.com>
Signed-off-by: Gao Xiang <xiang@kernel.org>